OpenAPI: Application creation based on Application object
Implemented Application creation OpenAPI based on new Application object
To #913
